Revert "docs: migrate search to Algolia (#3452)"
This reverts commit 7bd3e0669cebdc142a302e5423a1b31bd3b76504.
